修改：优化并重写部分高频调用方法为 StringUtil 类

+++ src/main/java/com/now/nowbot/model/osu/Beatmap.kt
@@ -124,7 +124,7 @@ data class Beatmap(
     var playCount: Int = 0,
 
     @field:JsonProperty("ranked")
-    var ranked: Int = 0,
+    var ranked: Byte = 0,
 
     @field:JsonProperty("url")
     var url: String? = null,
